JobSpy-v2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
sqlalchemy-utils==0.41.1
alembic==1.12.1
asyncpg==0.29.0
PyJWT==2.8.0
email-validator==2.1.0
aiohttp==3.9.1
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
openai==1.3.7
pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
playwright==1.40.0
minio==7.2.0